README.md
In the cooking and food industry, it's tricky to always make really good food. People want food to taste great every time they eat it. To do this, we need to find new ways beyond the usual ways of cooking. This is where machine learning comes in. It's like using smart technology to predict and make sure the food turns out amazing every time.